// yiqiheshebei/src/service/equ-manage/equ-manage/types.ts
export interface RequestDataLabName {
  data: {
    id: number
    title: string
  }[]
}

// 实验室批量配置保存
export interface StoreItemType {
  id: number
  number_plies: number
  is_del: number
  numbers: string
  qr_code: string
  school_id: number
  school_laboratory_id: number
  school_laboratory_name: string
  subject_id: null
  title: string
  type: number
  user_id: number
  user_name: string
  url: string
}

export const RequestDataLab: RequestDataLabName = {
  data: [
    {
      id: 1,
      title: ''
    }
  ]
}

//库存列表
export interface equList {
  inventory: string //
  lend: string //
  maintenance: string //
  scrap: string //
  total: number //数据总条数
  data: {
    id: number
    number: string //分类编号
    school_id: number // 学校id
    asset_number: string //学校编号(资产编号)
    name: string // 仪器名称
    specification: string //规格型号
    unit: string //单位
    total: number //总库存
    user_num: number //可用数量
    lend_num: number //出街数量
    maintenance_num: number //维修保养数量
    scrap_num: number //损坏报废数量
    category: [] //分类
    cost_type: number //价值分类
    type: number //1仪器2设备
  }[]
}
// ...
export interface allotList {
  school_instrument_id: number
  number: string
  asset_number: string
  name: string
  specification: string
  mode: string
  bacth: string
  type: number
  list: {
    type: string
    school_laboratory_id: number
    store_id: number
    number_plies: number
    total: number
  }[]
  numbers: string
}
// ...
export interface WarehousingList {
  school_instrument_id: number
  school_instrument_name: string
  unit: string
  number: string
  asset_number: string
  name: string
  specification: string
  type: number
  location: {
    school_laboratory_id: number | string
    store_id: number | string
    number_plies: number | string
    total: number
  }[]
  source: string
  voucher_number: string
  bacth: string | number
  factory: string
  unit_price: string
}
// ...
export interface detailList {
  id: number
  number: string
  asset_number: string
  specification: string
  name: string
  info: string
  unit: string
  number_year: string
  number_year_unit: string
  norm: string
  total: number
  cost_type: number
  picture: {
    url_path: string
  }[]
  is_must: number // 配备类型  1必配2选配
  is_group: number // 实验类型 1分组2演示3两者皆可
  sourceList: {
    source: string
    voucher_number: string
    bacth: string
    factory: string
    total: number
    qr_code: string
  }[]
  locationList: {
    store_name: string
    school_laboratory_name: string
    number_plies: number
    total: number
  }[]
}
// ...
export interface checkList {
  school_instrument_id: number
  instrument_id: number
  number: string
  asset_number: string
  name: string
  type: number
  specification: string
  yk: number
  bacth: string
  school_laboratory_id: number
  store_id: number
  number_plies: number
  num: number
  pd_desc: string
  pd_time: string
}